feat(mac): 新增 macOS 版工程并跑通(Xcode 26 + Rosetta x86_64)
在本机成功编译并运行 NarutoSenki macOS 版:

工程
- 新增 projects/NarutoSenki/proj.mac/(基于官方 multi-platform-cpp 模板重建,
  HelloCpp→NarutoSenki),纳入全部 Classes/*.cpp + MyUtils(含 sqlite3.c),
  内联编译 extensions 并链接 libcocos2dx.a
- Resources 以文件夹引用整体入包;Info.plist 更新 bundleId=
  dev.leaderone.narutosenki、版本 1.23(24),窗口标题「火影战记 NarutoSenki」

现代 SDK 兼容(CocosDenshion/mac)
- CDXMacOSXSupport.h:注释掉与现代 AudioToolbox 冲突的 AudioSession 枚举重定义
- CDAudioManager.m:isOtherAudioPlaying / isDeviceMuted 改为 macOS 安全实现
  (AudioSession* 系 API 在现代 macOS 已不可用,均为 iOS 专属)

64 位适配 / 启动修复
- ActionManager.cpp:7 处 (int)date 指针→int 截断改 (int)(intptr_t)date
  (原代码假定 32 位,x86_64 下报错;字节安全改动,GBK 完好)
- AppDelegate.cpp:Mac 下追加 addSearchPath("Resources")(仅 Mac,不影响他平台)
- AppController.mm:run() 前显式 makeCurrentContext,修复 gatherGPUInfo()
  在无当前 GL 上下文时 glGetString 返回 NULL 导致 std::string(NULL) abort

验证:启动稳定,GPU 信息正确采集(Apple M4 / GL 2.1 Metal),
通过菜单+BGM 阶段无闪退、无新崩溃报告。

Entire-Checkpoint: ec4ac16de4cd

+++ projects/NarutoSenki/Classes/ActionManager.cpp
@@ -2459,7 +2459,7 @@ void ActionManager::setMove(CCNode* sender,void* date){
 		} 
 		moveLength=this->getKnockLength();
 	}else{
-		moveLength=(int) date;
+		moveLength=(int)(intptr_t) date;
 	}
 
 	if( this->getPositionX()>_delegate->currentMap->getTileSize().width &&
@@ -2514,7 +2514,7 @@ void ActionManager::setJump(CCNode* sender,void* date){
 
 
 void ActionManager::setCharge(CCNode* sender,void* date){
-	int moveLength=(int) date;
+	int moveLength=(int)(intptr_t) date;
 	if( (this->getPositionX()<_delegate->currentMap->getTileSize().width && _isFlipped) ||
 		(this->getPositionX()>(_delegate->currentMap->getMapSize().width-1)*_delegate->currentMap->getTileSize().width && !_isFlipped)
 		){
@@ -2529,7 +2529,7 @@ void ActionManager::setCharge(CCNode* sender,void* date){
 
 // with out getCollider
 void ActionManager::setChargeB(CCNode* sender,void* date){
-	int moveLength=(int) date;
+	int moveLength=(int)(intptr_t) date;
 	float delay;
 	if(_actionState==ACTION_STATE_OATTACK || _actionState==ACTION_STATE_O2ATTACK){
 		delay=0.4f;
@@ -2902,7 +2902,7 @@ void ActionManager::setCommand(CCNode* sender,void* date){
 
 
 void ActionManager::setBuff(CCNode* sender,void* date){
-	int buffValue=(int) date;
+	int buffValue=(int)(intptr_t) date;
 	float buffStayTime=_attackRangeY;
 
 	if(strcmp(_attackType->getCString(),"hBuff")==0){
@@ -5435,7 +5435,7 @@ void ActionManager::stopMove(float dt){
 }
 
 void ActionManager::stopJump(CCNode* sender,void* date){
-	int stopTime=(int) date;
+	int stopTime=(int)(intptr_t) date;
 
 	if(_actionState==ACTION_STATE_JUMP){
 		this->getActionManager()->pauseTarget(this);
@@ -5580,7 +5580,7 @@ void ActionManager::setBulletGroup(float dt){
 
 
 void ActionManager::setClone(CCNode* sender,void* date){
-	int cloneTime=(int) date;
+	int cloneTime=(int)(intptr_t) date;
 	if(!_monsterArray){ _monsterArray=CCArray::create(); _monsterArray->retain(); } /* ensure summon array exists before any addObject: fixes null-deref crash on clone/summon skills */
 	Hero* clone=Hero::create();
 	
@@ -6432,7 +6432,7 @@ void ActionManager::removeSelf(float dt){
 }
 
 void ActionManager::setMonAttack(CCNode* sender,void* date){
-	int skillNum=(int) date;
+	int skillNum=(int)(intptr_t) date;
 	if(this->getMonsterArray()){
 		CCObject* pObject;
 		CCARRAY_FOREACH(this->getMonsterArray(),pObject){
